creating user audit trail and some renaming
